Can successfully import the hasarm function from another python file
