+++ spot_driver/launch/spot_image_publishers.launch.py
@@ -1,19 +1,17 @@
 # Copyright (c) 2023-2024 Boston Dynamics AI Institute LLC. All rights reserved.
 
-import logging
 import os
 from enum import Enum
-from typing import List, Optional, Tuple
+from typing import List
 
 import launch
 import launch_ros
-import yaml
 from launch import LaunchContext, LaunchDescription
 from launch.actions import DeclareLaunchArgument, OpaqueFunction
 from launch.conditions import IfCondition
 from launch.substitutions import LaunchConfiguration, PathJoinSubstitution
 
-from spot_wrapper.wrapper import SpotWrapper
+from spot_driver.launch.spot_launch_helpers import spot_has_arm
 
 
 class DepthRegisteredMode(Enum):
@@ -99,64 +97,6 @@ def create_point_cloud_nodelets(
             ),
         )
     return composable_node_descriptions
-
-
-def get_login_parameters(context: LaunchContext) -> Tuple[str, str, str, Optional[int], Optional[str]]:
-    """Obtain the username, password, hostname, and port of Spot from the environment variables or, if they are not
-    set, the configuration file yaml."""
-    # Get value from environment variables
-    username = os.getenv("BOSDYN_CLIENT_USERNAME")
-    password = os.getenv("BOSDYN_CLIENT_PASSWORD")
-    hostname = os.getenv("SPOT_IP")
-    portnum = os.getenv("SPOT_PORT")
-    port = int(portnum) if portnum else None
-    certificate = os.getenv("SPOT_CERTIFICATE")
-
-    # parse the yaml to determine if login information is set there
-    config_file_path = LaunchConfiguration("config_file").perform(context)
-    if os.path.isfile(config_file_path):
-        with open(config_file_path, "r") as config_yaml:
-            try:
-                config_dict = yaml.safe_load(config_yaml)
-                if ("/**" in config_dict) and ("ros__parameters" in config_dict["/**"]):
-                    ros_params = config_dict["/**"]["ros__parameters"]
-                    # only set username/password/hostname if they were not already set as environment variables.
-                    if (not username) and ("username" in ros_params):
-                        username = ros_params["username"]
-                    if (not password) and ("password" in ros_params):
-                        password = ros_params["password"]
-                    if (not hostname) and ("hostname" in ros_params):
-                        hostname = ros_params["hostname"]
-                    if not port and "port" in ros_params:
-                        port = ros_params["port"]
-                    if not certificate and "certificate" in ros_params:
-                        certificate = ros_params["certificate"]
-            except yaml.YAMLError as exc:
-                print("Parsing config_file yaml failed with: {}".format(exc))
-    if (not username) or (not password) or (not hostname):
-        raise ValueError(
-            "One or more of your login credentials has not been specified! Got invalid values of "
-            "[Username: '{}' Password: '{}' Hostname: '{}']. Ensure that your environment variables are set or "
-            "update your config_file yaml.".format(username, password, hostname)
-        )
-    return username, password, hostname, port, certificate
-
-
-def spot_has_arm(context: LaunchContext) -> bool:
-    """Check if Spot has an arm by logging in and instantiating a SpotWrapper"""
-    spot_name = LaunchConfiguration("spot_name").perform(context)
-    logger = logging.getLogger("spot_driver_launch")
-    username, password, hostname, port, certificate = get_login_parameters(context)
-    spot_wrapper = SpotWrapper(
-        username=username,
-        password=password,
-        hostname=hostname,
-        port=port,
-        cert_resource_glob=certificate,
-        robot_name=spot_name,
-        logger=logger,
-    )
-    return spot_wrapper.has_arm()
 
 
 def launch_setup(context: LaunchContext, ld: LaunchDescription) -> None:
